🤖 refactor(site/src/pages): nest child routes
