+++ site/src/pages/TasksPage/TaskPage/TaskTopbar.tsx
@@ -21,7 +21,7 @@ import {
 	TooltipTrigger,
 } from "#/components/Tooltip/Tooltip";
 import { useClipboard } from "#/hooks/useClipboard";
-import { ShareButton } from "../WorkspacePage/WorkspaceActions/ShareButton";
+import { ShareButton } from "#/pages/WorkspacePage/WorkspaceActions/ShareButton";
 import { TaskStartupWarningButton } from "./TaskStartupWarningButton";
 import { TaskStatusLink } from "./TaskStatusLink";
