+++ site/src/pages/WorkspacePage/TerminalPage/TerminalCommandConsentDialog.tsx
