+++ site/src/pages/WorkspacePage/TerminalPage/TerminalPage.test.tsx
