+++ site/src/pages/WorkspacePage/TerminalPage/TerminalPage.tsx
